Refactored some components to separate files
